Moved Implementation code to internal
Made GridFSBuckets depend on MongoDatabaseImpl
