+++ driver/src/main/com/mongodb/reactivestreams/client/gridfs/helpers/AsyncStreamHelper.java
@@ -16,24 +16,14 @@
 
 package com.mongodb.reactivestreams.client.gridfs.helpers;
 
-import com.mongodb.Block;
-import com.mongodb.MongoGridFSException;
-import com.mongodb.async.SingleResultCallback;
-import com.mongodb.reactivestreams.client.Success;
 import com.mongodb.reactivestreams.client.gridfs.AsyncInputStream;
 import com.mongodb.reactivestreams.client.gridfs.AsyncOutputStream;
-import com.mongodb.reactivestreams.client.internal.ObservableToPublisher;
-import org.reactivestreams.Publisher;
-import org.reactivestreams.Subscriber;
+import com.mongodb.reactivestreams.client.internal.GridFSAsyncStreamHelper;
 
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.ByteBuffer;
 
-import static com.mongodb.async.client.Observables.observe;
-import static com.mongodb.reactivestreams.client.internal.GridFSHelper.toCallbackAsyncInputStream;
-import static com.mongodb.reactivestreams.client.internal.PublisherHelper.voidToSuccessCallback;
-import static org.bson.assertions.Assertions.notNull;
 
 /**
  * A general helper class that creates {@link AsyncInputStream} or {@link AsyncOutputStream} instances.
@@ -57,7 +47,8 @@ public final class AsyncStreamHelper {
      * @return the AsyncInputStream
      */
     public static AsyncInputStream toAsyncInputStream(final byte[] srcBytes) {
-        return toAsyncInputStream(ByteBuffer.wrap(srcBytes));
+        return GridFSAsyncStreamHelper.toAsyncInputStream(com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
+                .toAsyncInputStream(srcBytes));
     }
 
     /**
@@ -67,7 +58,8 @@ public final class AsyncStreamHelper {
      * @return the AsyncOutputStream
      */
     public static AsyncOutputStream toAsyncOutputStream(final byte[] dstBytes) {
-        return toAsyncOutputStream(ByteBuffer.wrap(dstBytes));
+        return GridFSAsyncStreamHelper.toAsyncOutputStream(com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
+                .toAsyncOutputStream(dstBytes));
     }
 
     /**
@@ -77,29 +69,8 @@ public final class AsyncStreamHelper {
      * @return the AsyncInputStream
      */
     public static AsyncInputStream toAsyncInputStream(final ByteBuffer srcByteBuffer) {
-        final com.mongodb.async.client.gridfs.AsyncInputStream wrapper = com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
-                .toAsyncInputStream(srcByteBuffer);
-        return new AsyncInputStream() {
-            @Override
-            public Publisher<Integer> read(final ByteBuffer dstByteBuffer) {
-                return new ObservableToPublisher<Integer>(observe(new Block<SingleResultCallback<Integer>>() {
-                    @Override
-                    public void apply(final SingleResultCallback<Integer> callback) {
-                        wrapper.read(srcByteBuffer, callback);
-                    }
-                }));
-            }
-
-            @Override
-            public Publisher<Success> close() {
-                return new ObservableToPublisher<Success>(observe(new Block<SingleResultCallback<Success>>() {
-                    @Override
-                    public void apply(final SingleResultCallback<Success> callback) {
-                        wrapper.close(voidToSuccessCallback(callback));
-                    }
-                }));
-            }
-        };
+        return GridFSAsyncStreamHelper.toAsyncInputStream(com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
+                .toAsyncInputStream(srcByteBuffer));
     }
 
     /**
@@ -109,29 +80,8 @@ public final class AsyncStreamHelper {
      * @return the AsyncOutputStream
      */
     public static AsyncOutputStream toAsyncOutputStream(final ByteBuffer dstByteBuffer) {
-        final com.mongodb.async.client.gridfs.AsyncOutputStream wrapper = com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
-                .toAsyncOutputStream(dstByteBuffer);
-        return new AsyncOutputStream() {
-            @Override
-            public Publisher<Integer> write(final ByteBuffer src) {
-                return new ObservableToPublisher<Integer>(observe(new Block<SingleResultCallback<Integer>>() {
-                    @Override
-                    public void apply(final SingleResultCallback<Integer> callback) {
-                        wrapper.write(src, callback);
-                    }
-                }));
-            }
-
-            @Override
-            public Publisher<Success> close() {
-                return new ObservableToPublisher<Success>(observe(new Block<SingleResultCallback<Success>>() {
-                    @Override
-                    public void apply(final SingleResultCallback<Success> callback) {
-                        wrapper.close(voidToSuccessCallback(callback));
-                    }
-                }));
-            }
-        };
+        return GridFSAsyncStreamHelper.toAsyncOutputStream(com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
+                .toAsyncOutputStream(dstByteBuffer));
     }
 
     /**
@@ -141,30 +91,8 @@ public final class AsyncStreamHelper {
      * @return the AsyncInputStream
      */
     public static AsyncInputStream toAsyncInputStream(final InputStream inputStream) {
-        final com.mongodb.async.client.gridfs.AsyncInputStream wrapper = com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
-                .toAsyncInputStream(inputStream);
-
-        return new AsyncInputStream() {
-            @Override
-            public Publisher<Integer> read(final ByteBuffer dst) {
-                return new ObservableToPublisher<Integer>(observe(new Block<SingleResultCallback<Integer>>() {
-                    @Override
-                    public void apply(final SingleResultCallback<Integer> callback) {
-                        wrapper.read(dst, callback);
-                    }
-                }));
-            }
-
-            @Override
-            public Publisher<Success> close() {
-                return new ObservableToPublisher<Success>(observe(new Block<SingleResultCallback<Success>>() {
-                    @Override
-                    public void apply(final SingleResultCallback<Success> callback) {
-                        wrapper.close(voidToSuccessCallback(callback));
-                    }
-                }));
-            }
-        };
+        return GridFSAsyncStreamHelper.toAsyncInputStream(com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
+                .toAsyncInputStream(inputStream));
     }
 
     /**
@@ -174,30 +102,8 @@ public final class AsyncStreamHelper {
      * @return the AsyncOutputStream
      */
     public static AsyncOutputStream toAsyncOutputStream(final OutputStream outputStream) {
-        final com.mongodb.async.client.gridfs.AsyncOutputStream wrapper = com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
-                .toAsyncOutputStream(outputStream);
-        return new AsyncOutputStream() {
-
-            @Override
-            public Publisher<Integer> write(final ByteBuffer src) {
-                return new ObservableToPublisher<Integer>(observe(new Block<SingleResultCallback<Integer>>() {
-                    @Override
-                    public void apply(final SingleResultCallback<Integer> callback) {
-                        wrapper.write(src, callback);
-                    }
-                }));
-            }
-
-            @Override
-            public Publisher<Success> close() {
-                return new ObservableToPublisher<Success>(observe(new Block<SingleResultCallback<Success>>() {
-                    @Override
-                    public void apply(final SingleResultCallback<Success> callback) {
-                        wrapper.close(voidToSuccessCallback(callback));
-                    }
-                }));
-            }
-        };
+        return GridFSAsyncStreamHelper.toAsyncOutputStream(com.mongodb.async.client.gridfs.helpers.AsyncStreamHelper
+                .toAsyncOutputStream(outputStream));
     }
 
     private AsyncStreamHelper() {
