+++ driver/src/main/com/mongodb/reactivestreams/client/internal/FindPublisherImpl.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package com.mongodb.reactivestreams.client;
+package com.mongodb.reactivestreams.client.internal;
 
 import com.mongodb.Block;
 import com.mongodb.CursorType;
 import com.mongodb.async.SingleResultCallback;
 import com.mongodb.client.model.Collation;
-import com.mongodb.reactivestreams.client.internal.ObservableToPublisher;
+import com.mongodb.reactivestreams.client.FindPublisher;
 import org.bson.conversions.Bson;
 import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
@@ -30,7 +30,8 @@ import java.util.concurrent.TimeUnit;
 import static com.mongodb.assertions.Assertions.notNull;
 import static com.mongodb.async.client.Observables.observe;
 
-class FindPublisherImpl<TResult> implements FindPublisher<TResult> {
+
+final class FindPublisherImpl<TResult> implements FindPublisher<TResult> {
 
     private final com.mongodb.async.client.FindIterable<TResult> wrapped;
