+++ driver/src/main/com/mongodb/reactivestreams/client/internal/GridFSAsyncStreamHelper.java
@@ -1,17 +1,90 @@
 package com.mongodb.reactivestreams.client.internal;
 
+import com.mongodb.Block;
 import com.mongodb.async.SingleResultCallback;
 import com.mongodb.reactivestreams.client.Success;
 import com.mongodb.reactivestreams.client.gridfs.AsyncInputStream;
 import com.mongodb.reactivestreams.client.gridfs.AsyncOutputStream;
+import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
 import org.reactivestreams.Subscription;
 
 import java.nio.ByteBuffer;
 
-public final class GridFSHelper {
+import static com.mongodb.async.client.Observables.observe;
+import static com.mongodb.reactivestreams.client.internal.PublisherHelper.voidToSuccessCallback;
 
-    public static com.mongodb.async.client.gridfs.AsyncInputStream toCallbackAsyncInputStream(final AsyncInputStream wrapped) {
+/**
+ * Internal GridFS AsyncStream Helper
+ *
+ * <p>This should not be considered a part of the public API.</p>
+ */
+public final class GridFSAsyncStreamHelper {
+
+    /**
+     * Converts the callback AsyncInputStream to a Publisher AsyncInputStream
+     *
+     * <p>This should not be considered a part of the public API.</p>
+     * @param wrapper the callback AsyncInputStream
+     * @return the Publisher AsyncInputStream
+     */
+    public static AsyncInputStream toAsyncInputStream(final com.mongodb.async.client.gridfs.AsyncInputStream wrapper) {
+        return new AsyncInputStream() {
+            @Override
+            public Publisher<Integer> read(final ByteBuffer dst) {
+                return new ObservableToPublisher<Integer>(observe(new Block<SingleResultCallback<Integer>>() {
+                    @Override
+                    public void apply(final SingleResultCallback<Integer> callback) {
+                        wrapper.read(dst, callback);
+                    }
+                }));
+            }
+
+            @Override
+            public Publisher<Success> close() {
+                return new ObservableToPublisher<Success>(observe(new Block<SingleResultCallback<Success>>() {
+                    @Override
+                    public void apply(final SingleResultCallback<Success> callback) {
+                        wrapper.close(voidToSuccessCallback(callback));
+                    }
+                }));
+            }
+        };
+    }
+
+    /**
+     * Converts the callback AsyncOutputStream to a Publisher AsyncOutputStream
+     *
+     * <p>This should not be considered a part of the public API.</p>
+     * @param wrapper the callback AsyncOutputStream
+     * @return the Publisher AsyncOutputStream
+     */
+    public static AsyncOutputStream toAsyncOutputStream(final com.mongodb.async.client.gridfs.AsyncOutputStream wrapper) {
+        return new AsyncOutputStream() {
+
+            @Override
+            public Publisher<Integer> write(final ByteBuffer src) {
+                return new ObservableToPublisher<Integer>(observe(new Block<SingleResultCallback<Integer>>() {
+                    @Override
+                    public void apply(final SingleResultCallback<Integer> callback) {
+                        wrapper.write(src, callback);
+                    }
+                }));
+            }
+
+            @Override
+            public Publisher<Success> close() {
+                return new ObservableToPublisher<Success>(observe(new Block<SingleResultCallback<Success>>() {
+                    @Override
+                    public void apply(final SingleResultCallback<Success> callback) {
+                        wrapper.close(voidToSuccessCallback(callback));
+                    }
+                }));
+            }
+        };
+    }
+
+    static com.mongodb.async.client.gridfs.AsyncInputStream toCallbackAsyncInputStream(final AsyncInputStream wrapped) {
         return new com.mongodb.async.client.gridfs.AsyncInputStream() {
 
             @Override
@@ -70,7 +143,7 @@ public final class GridFSHelper {
         };
     }
 
-    public static com.mongodb.async.client.gridfs.AsyncOutputStream toCallbackAsyncOutputStream(final AsyncOutputStream wrapped) {
+    static com.mongodb.async.client.gridfs.AsyncOutputStream toCallbackAsyncOutputStream(final AsyncOutputStream wrapped) {
         return new com.mongodb.async.client.gridfs.AsyncOutputStream() {
 
             @Override
@@ -129,6 +202,6 @@ public final class GridFSHelper {
         };
     }
 
-    private GridFSHelper() {
+    private GridFSAsyncStreamHelper() {
     }
 }
