+++ driver/src/main/com/mongodb/reactivestreams/client/internal/ListIndexesPublisherImpl.java
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package com.mongodb.reactivestreams.client;
+package com.mongodb.reactivestreams.client.internal;
 
-import com.mongodb.client.model.Collation;
-import com.mongodb.reactivestreams.client.internal.ObservableToPublisher;
-import org.bson.conversions.Bson;
+import com.mongodb.reactivestreams.client.ListIndexesPublisher;
 import org.reactivestreams.Subscriber;
 
 import java.util.concurrent.TimeUnit;
@@ -26,30 +24,18 @@ import java.util.concurrent.TimeUnit;
 import static com.mongodb.assertions.Assertions.notNull;
 import static com.mongodb.async.client.Observables.observe;
 
+final class ListIndexesPublisherImpl<TResult> implements ListIndexesPublisher<TResult> {
 
-class DistinctPublisherImpl<TResult> implements DistinctPublisher<TResult> {
+    private final com.mongodb.async.client.ListIndexesIterable<TResult> wrapped;
 
-    private final com.mongodb.async.client.DistinctIterable<TResult> wrapped;
-
-    DistinctPublisherImpl(final com.mongodb.async.client.DistinctIterable<TResult> wrapped) {
+    ListIndexesPublisherImpl(final com.mongodb.async.client.ListIndexesIterable<TResult> wrapped) {
         this.wrapped = notNull("wrapped", wrapped);
     }
 
     @Override
-    public DistinctPublisher<TResult> filter(final Bson filter) {
-        wrapped.filter(filter);
-        return this;
-    }
-
-    @Override
-    public DistinctPublisher<TResult> maxTime(final long maxTime, final TimeUnit timeUnit) {
+    public ListIndexesPublisherImpl<TResult> maxTime(final long maxTime, final TimeUnit timeUnit) {
+        notNull("timeUnit", timeUnit);
         wrapped.maxTime(maxTime, timeUnit);
-        return this;
-    }
-
-    @Override
-    public DistinctPublisher<TResult> collation(final Collation collation) {
-        wrapped.collation(collation);
         return this;
     }
 
@@ -58,3 +44,4 @@ class DistinctPublisherImpl<TResult> implements DistinctPublisher<TResult> {
         new ObservableToPublisher<TResult>(observe(wrapped)).subscribe(s);
     }
 }
+
