+++ driver/src/main/com/mongodb/reactivestreams/client/internal/MapReducePublisherImpl.java
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package com.mongodb.reactivestreams.client;
+package com.mongodb.reactivestreams.client.internal;
 
 import com.mongodb.Block;
 import com.mongodb.async.SingleResultCallback;
 import com.mongodb.client.model.Collation;
 import com.mongodb.client.model.MapReduceAction;
-import com.mongodb.reactivestreams.client.internal.ObservableToPublisher;
+import com.mongodb.reactivestreams.client.MapReducePublisher;
+import com.mongodb.reactivestreams.client.Success;
 import org.bson.conversions.Bson;
 import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
@@ -31,8 +32,7 @@ import static com.mongodb.assertions.Assertions.notNull;
 import static com.mongodb.async.client.Observables.observe;
 import static com.mongodb.reactivestreams.client.internal.PublisherHelper.voidToSuccessCallback;
 
-
-class MapReducePublisherImpl<TResult> implements MapReducePublisher<TResult> {
+final class MapReducePublisherImpl<TResult> implements MapReducePublisher<TResult> {
 
 
     private final com.mongodb.async.client.MapReduceIterable<TResult> wrapped;
