+++ driver/src/main/com/mongodb/reactivestreams/client/internal/MongoDatabaseImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.reactivestreams.client;
+package com.mongodb.reactivestreams.client.internal;
 
 import com.mongodb.Block;
 import com.mongodb.ReadConcern;
@@ -23,7 +23,10 @@ import com.mongodb.WriteConcern;
 import com.mongodb.async.SingleResultCallback;
 import com.mongodb.client.model.CreateCollectionOptions;
 import com.mongodb.client.model.CreateViewOptions;
-import com.mongodb.reactivestreams.client.internal.ObservableToPublisher;
+import com.mongodb.reactivestreams.client.ListCollectionsPublisher;
+import com.mongodb.reactivestreams.client.MongoCollection;
+import com.mongodb.reactivestreams.client.MongoDatabase;
+import com.mongodb.reactivestreams.client.Success;
 import org.bson.Document;
 import org.bson.codecs.configuration.CodecRegistry;
 import org.bson.conversions.Bson;
@@ -35,7 +38,12 @@ import static com.mongodb.assertions.Assertions.notNull;
 import static com.mongodb.async.client.Observables.observe;
 import static com.mongodb.reactivestreams.client.internal.PublisherHelper.voidToSuccessCallback;
 
-class MongoDatabaseImpl implements MongoDatabase {
+/**
+ * The internal MongoDatabase implementation.
+ *
+ * <p>This should not be considered a part of the public API.</p>
+ */
+public class MongoDatabaseImpl implements MongoDatabase {
 
     private final com.mongodb.async.client.MongoDatabase wrapped;
 
@@ -190,8 +198,13 @@ class MongoDatabaseImpl implements MongoDatabase {
         }));
     }
 
-    @Override
-    public com.mongodb.async.client.MongoDatabase getAsyncMongoDatabase() {
+    /**
+     * Gets the wrapped MongoDatabase
+     *
+     * <p>This should not be considered a part of the public API.</p>
+     * @return wrapped MongoDatabase
+     */
+    public com.mongodb.async.client.MongoDatabase getWrapped() {
         return wrapped;
     }
 }
