+++ driver/src/test/unit/com/mongodb/reactivestreams/client/ListIndexesPublisherSpecification.groovy
@@ -16,67 +16,20 @@
 
 package com.mongodb.reactivestreams.client
 
-import com.mongodb.MongoNamespace
-import com.mongodb.async.client.ListIndexesIterableImpl
+import com.mongodb.async.client.ListIndexesIterable
 import com.mongodb.async.client.MongoIterable
-import com.mongodb.operation.ListIndexesOperation
-import org.bson.Document
-import org.bson.codecs.BsonValueCodecProvider
-import org.bson.codecs.DocumentCodec
-import org.bson.codecs.DocumentCodecProvider
-import org.bson.codecs.ValueCodecProvider
 import org.reactivestreams.Publisher
-import org.reactivestreams.Subscriber
 import spock.lang.Specification
-
-import static com.mongodb.ReadPreference.secondary
-import static com.mongodb.reactivestreams.client.CustomMatchers.isTheSameAs
-import static java.util.concurrent.TimeUnit.MILLISECONDS
-import static org.bson.codecs.configuration.CodecRegistries.fromProviders
-import static spock.util.matcher.HamcrestSupport.expect
 
 class ListIndexesPublisherSpecification extends Specification {
 
     def 'should have the same methods as the wrapped ListIndexesIterable'() {
         given:
-        def wrapped = (com.mongodb.async.client.ListIndexesIterable.methods*.name - MongoIterable.methods*.name).sort()
+        def wrapped = (ListIndexesIterable.methods*.name - MongoIterable.methods*.name).sort()
         def local = (ListIndexesPublisher.methods*.name - Publisher.methods*.name - 'batchSize').sort()
 
         expect:
         wrapped == local
-    }
-
-    def 'should build the expected listIndexesOperation'() {
-        given:
-        def subscriber = Stub(Subscriber) {
-            onSubscribe(_) >> { args -> args[0].request(100) }
-        }
-        def namespace = new MongoNamespace('db', 'coll')
-        def codecRegistry = fromProviders([new DocumentCodecProvider(), new BsonValueCodecProvider(), new ValueCodecProvider()])
-        def executor = new TestOperationExecutor([null, null]);
-        def wrapped = new ListIndexesIterableImpl(namespace, Document, codecRegistry, secondary(), executor)
-        def listIndexesPublisher = new ListIndexesPublisherImpl<Document>(wrapped)
-
-        when: 'default input should be as expected'
-        listIndexesPublisher.subscribe(subscriber)
-
-        def operation = executor.getReadOperation() as ListIndexesOperation<Document>
-        def readPreference = executor.getReadPreference()
-
-        then:
-        expect operation, isTheSameAs(new ListIndexesOperation<Document>(namespace, new DocumentCodec()).batchSize(100))
-        readPreference == secondary()
-
-        when: 'overriding initial options'
-        listIndexesPublisher
-                .maxTime(999, MILLISECONDS)
-                .subscribe(subscriber)
-
-        operation = executor.getReadOperation() as ListIndexesOperation<Document>
-
-        then: 'should use the overrides'
-        expect operation, isTheSameAs(new ListIndexesOperation<Document>(namespace, new DocumentCodec())
-                .batchSize(100).maxTime(999, MILLISECONDS))
     }
 
 }
