+++ driver/src/test/unit/com/mongodb/reactivestreams/client/MongoDatabaseSpecification.groovy
@@ -16,21 +16,8 @@
 
 package com.mongodb.reactivestreams.client
 
-import com.mongodb.ReadConcern
-import com.mongodb.ReadPreference
-import com.mongodb.WriteConcern
-import com.mongodb.async.client.MongoCollection as WrappedMongoCollection
 import com.mongodb.async.client.MongoDatabase as WrappedMongoDatabase
-import com.mongodb.client.model.CreateCollectionOptions
-import com.mongodb.client.model.CreateViewOptions
-import org.bson.BsonDocument
-import org.bson.Document
-import org.bson.codecs.configuration.CodecRegistry
-import org.reactivestreams.Subscriber
 import spock.lang.Specification
-
-import static com.mongodb.reactivestreams.client.CustomMatchers.isTheSameAs
-import static spock.util.matcher.HamcrestSupport.expect
 
 class MongoDatabaseSpecification extends Specification {
 
@@ -41,319 +28,5 @@ class MongoDatabaseSpecification extends Specification {
 
         expect:
         wrapped == local
-    }
-
-    def 'should return the a collection'() {
-        given:
-        def wrappedCollection = Mock(WrappedMongoCollection)
-        def wrapped = Mock(WrappedMongoDatabase) {
-            getCollection(_) >> wrappedCollection
-            getCollection(_, _) >> wrappedCollection
-        }
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        def collection = mongoDatabase.getCollection('collectionName')
-
-        then:
-        expect collection, isTheSameAs(new MongoCollectionImpl(wrappedCollection))
-
-        when:
-        collection = mongoDatabase.getCollection('collectionName', Document)
-
-        then:
-        expect collection, isTheSameAs(new MongoCollectionImpl(wrappedCollection))
-    }
-
-    def 'should call the underlying getName'() {
-        given:
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.getName()
-
-        then:
-        1 * wrapped.getName()
-    }
-
-    def 'should call the underlying getCodecRegistry'() {
-        given:
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.getCodecRegistry()
-
-        then:
-        1 * wrapped.getCodecRegistry()
-    }
-    def 'should call the underlying getReadPreference'() {
-        given:
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.getReadPreference()
-
-        then:
-        1 * wrapped.getReadPreference()
-
-    }
-
-    def 'should call the underlying getWriteConcern'() {
-        given:
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.getWriteConcern()
-
-        then:
-        1 * wrapped.getWriteConcern()
-    }
-
-    def 'should call the underlying getReadConcern'() {
-        given:
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.getReadConcern()
-
-        then:
-        1 * wrapped.getReadConcern()
-    }
-
-    def 'should call the underlying withCodecRegistry'() {
-        given:
-        def codecRegistry = Stub(CodecRegistry)
-        def wrappedResult = Stub(WrappedMongoDatabase)
-        def wrapped = Mock(WrappedMongoDatabase) {
-            1 * withCodecRegistry(codecRegistry) >> wrappedResult
-        }
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        def result = mongoDatabase.withCodecRegistry(codecRegistry)
-
-        then:
-        expect result, isTheSameAs(new MongoDatabaseImpl(wrappedResult))
-    }
-
-    def 'should call the underlying withReadPreference'() {
-        given:
-        def readPreference = Stub(ReadPreference)
-        def wrappedResult = Stub(WrappedMongoDatabase)
-        def wrapped = Mock(WrappedMongoDatabase) {
-            1 * withReadPreference(readPreference) >> wrappedResult
-        }
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        def result = mongoDatabase.withReadPreference(readPreference)
-
-        then:
-        expect result, isTheSameAs(new MongoDatabaseImpl(wrappedResult))
-    }
-
-    def 'should call the underlying withWriteConcern'() {
-        given:
-        def writeConcern = Stub(WriteConcern)
-        def wrappedResult = Stub(WrappedMongoDatabase)
-        def wrapped = Mock(WrappedMongoDatabase) {
-            1 * withWriteConcern(writeConcern) >> wrappedResult
-        }
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        def result = mongoDatabase.withWriteConcern(writeConcern)
-
-        then:
-        expect result, isTheSameAs(new MongoDatabaseImpl(wrappedResult))
-    }
-
-
-    def 'should call the underlying withReadConcern'() {
-        given:
-        def readConcern = ReadConcern.DEFAULT
-        def wrappedResult = Stub(WrappedMongoDatabase)
-        def wrapped = Mock(WrappedMongoDatabase) {
-            1 * withReadConcern(readConcern) >> wrappedResult
-        }
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        def result = mongoDatabase.withReadConcern(readConcern)
-
-        then:
-        expect result, isTheSameAs(new MongoDatabaseImpl(wrappedResult))
-    }
-
-    def 'should call the underlying runCommand when writing'() {
-        given:
-        def subscriber = Stub(Subscriber) {
-            onSubscribe(_) >> { args -> args[0].request(1) }
-        }
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.runCommand(new Document())
-
-        then: 'only executed when requested'
-        0 * wrapped.runCommand(_, _, _)
-
-        when:
-        mongoDatabase.runCommand(new Document()).subscribe(subscriber)
-
-        then:
-        1 * wrapped.runCommand(new Document(), Document, _)
-
-        when:
-        mongoDatabase.runCommand(new BsonDocument(), BsonDocument).subscribe(subscriber)
-
-        then:
-        1 * wrapped.runCommand(new BsonDocument(), BsonDocument, _)
-    }
-    def 'should call the underlying runCommand for read operations'() {
-        given:
-        def subscriber = Stub(Subscriber) {
-            onSubscribe(_) >> { args -> args[0].request(1) }
-        }
-        def readPreference = Stub(ReadPreference)
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.runCommand(new Document(), readPreference)
-
-        then: 'only executed when requested'
-        0 * wrapped.runCommand(_, _, _, _)
-
-        when:
-        mongoDatabase.runCommand(new Document(), readPreference).subscribe(subscriber)
-
-        then:
-        1 * wrapped.runCommand(new Document(), readPreference, Document, _)
-
-        when:
-        mongoDatabase.runCommand(new BsonDocument(), readPreference, BsonDocument).subscribe(subscriber)
-
-        then:
-        1 * wrapped.runCommand(new BsonDocument(), readPreference, BsonDocument, _)
-    }
-
-    def 'should call the underlying drop'() {
-        given:
-        def subscriber = Stub(Subscriber) {
-            onSubscribe(_) >> { args -> args[0].request(1) }
-        }
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.drop()
-
-        then: 'only executed when requested'
-        0 * wrapped.drop(_)
-
-        when:
-        mongoDatabase.drop().subscribe(subscriber)
-
-        then:
-        1 * wrapped.drop(_)
-    }
-    def 'should call the underlying listCollectionNames'() {
-        given:
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.listCollectionNames()
-
-        then:
-        1 * wrapped.listCollectionNames()
-
-    }
-    def 'should call the underlying listCollections'() {
-        given:
-        def wrappedResult = Stub(com.mongodb.async.client.ListCollectionsIterable)
-        def wrapped = Mock(WrappedMongoDatabase) {
-            1 * listCollections(Document) >> wrappedResult
-            1 * listCollections(BsonDocument) >> wrappedResult
-        }
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        def publisher = mongoDatabase.listCollections()
-
-        then:
-        expect publisher, isTheSameAs(new ListCollectionsPublisherImpl(wrappedResult))
-
-        when:
-        publisher = mongoDatabase.listCollections(BsonDocument)
-
-        then:
-        expect publisher, isTheSameAs(new ListCollectionsPublisherImpl(wrappedResult))
-    }
-
-    def 'should call the underlying createCollection'() {
-        given:
-        def subscriber = Stub(Subscriber) {
-            onSubscribe(_) >> { args -> args[0].request(1) }
-        }
-        def createCollectionOptions = Stub(CreateCollectionOptions)
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-
-        when:
-        mongoDatabase.createCollection('collectionName')
-
-        then: 'only executed when requested'
-        0 * wrapped.createCollection(_, _, _)
-
-        when:
-        mongoDatabase.createCollection('collectionName').subscribe(subscriber)
-
-        then:
-        1 * wrapped.createCollection('collectionName', _, _)
-
-        when:
-        mongoDatabase.createCollection('collectionName', createCollectionOptions).subscribe(subscriber)
-
-        then:
-        1 * wrapped.createCollection('collectionName', createCollectionOptions, _)
-    }
-
-    def 'should call the underlying createView'() {
-        given:
-        def subscriber = Stub(Subscriber) {
-            onSubscribe(_) >> { args -> args[0].request(1) }
-        }
-        def createViewOptions = Stub(CreateViewOptions)
-        def wrapped = Mock(WrappedMongoDatabase)
-        def mongoDatabase = new MongoDatabaseImpl(wrapped)
-        def viewName = 'view1'
-        def viewOn = 'col1'
-        def pipeline = [new Document('$match', new Document('x', true))];
-
-        when:
-        mongoDatabase.createView(viewName, viewOn, pipeline)
-
-        then: 'only executed when requested'
-        0 * wrapped.createView(_, _, _, _)
-
-        when:
-        mongoDatabase.createView(viewName, viewOn, pipeline).subscribe(subscriber)
-
-        then:
-        1 * wrapped.createView(viewName, viewOn, pipeline, _)
-
-        when:
-        mongoDatabase.createView(viewName, viewOn, pipeline, createViewOptions).subscribe(subscriber)
-
-        then:
-        1 * wrapped.createView(viewName, viewOn, pipeline, createViewOptions, _)
     }
 }
