+++ driver/src/test/unit/com/mongodb/reactivestreams/client/gridfs/GridFSDownloadStreamSpecification.groovy
@@ -16,68 +16,18 @@
 
 package com.mongodb.reactivestreams.client.gridfs
 
-import org.reactivestreams.Subscriber
 import spock.lang.Specification
-
-import java.nio.ByteBuffer
+import com.mongodb.async.client.gridfs.GridFSDownloadStream as WrappedGridFSDownloadStream
 
 class GridFSDownloadStreamSpecification extends Specification {
 
-    def subscriber = Stub(Subscriber) {
-        onSubscribe(_) >> { args -> args[0].request(100) }
-    }
-
     def 'should have the same methods as the wrapped GridFSDownloadStream'() {
         given:
-        def wrapped = (com.mongodb.async.client.gridfs.GridFSDownloadStream.methods*.name).sort()
+        def wrapped = (WrappedGridFSDownloadStream.methods*.name).sort()
         def local = (GridFSDownloadStream.methods*.name).sort()
 
         expect:
         wrapped == local
-    }
-
-    def 'should call the underlying getGridFSFile'() {
-        when:
-        def wrapped = Mock(com.mongodb.async.client.gridfs.GridFSDownloadStream) {
-            1 * getGridFSFile(_)
-        }
-        def downloadStream = new GridFSDownloadStreamImpl(wrapped)
-
-        then:
-        downloadStream.getGridFSFile().subscribe(subscriber)
-    }
-
-    def 'should call the underlying batchSize'() {
-        when:
-        def wrapped = Mock(com.mongodb.async.client.gridfs.GridFSDownloadStream) {
-            1 * batchSize(10)
-        }
-        def downloadStream = new GridFSDownloadStreamImpl(wrapped)
-
-        then:
-        downloadStream.batchSize(10)
-    }
-
-    def 'should call the underlying read'() {
-        when:
-        def wrapped = Mock(com.mongodb.async.client.gridfs.GridFSDownloadStream) {
-            1 * read(ByteBuffer.allocate(2), _)
-        }
-        def downloadStream = new GridFSDownloadStreamImpl(wrapped)
-
-        then:
-        downloadStream.read(ByteBuffer.allocate(2)).subscribe(subscriber)
-    }
-
-    def 'should call the underlying close'() {
-        when:
-        def wrapped = Mock(com.mongodb.async.client.gridfs.GridFSDownloadStream) {
-            1 * close(_)
-        }
-        def downloadStream = new GridFSDownloadStreamImpl(wrapped)
-
-        then:
-        downloadStream.close().subscribe(subscriber)
     }
 
 }
